:* init project
There are not well structured moments yet.

Error handling, I decided to left it in `.unwrap()` style in
order to get advantage in faster delop process.

Initial read timeout should be calculated as RTT * 2, but
I have no idea yet how to calculate RTT initially.

Connection to host. Original implementation of ping does a bit different thing what we do,
as I can say it make connection after reaching the remote host.
And may be it calculate RTT here, but it's not as easy as I'd like.
It's crusial factor actually since if you run two process in place
they will interfere without the connection.
Nevertheless I might be wrong about the connection. But this is the best idea I hold

Signed-off-by: Maxim Zhiburt <zhiburt@gmail.com>