packet: refactoring the trait
There's still incomplete field.
My consern here is build method which is
maybe be better have the old style.
Just to reuse the buffer.
But for example implementation of icmp header is held it
as Vec<u8> to modify it.
If we whan't to reduce allocations by removing the vector but still have the ability to
change the state of the header (increment seq) for example. we have to be able for modification.
But if we will take a &mut [u8] it will be not able to act in other operations...

Might we could separte the Packet trait to Parser + Builder...

Signed-off-by: Maxim Zhiburt <zhiburt@gmail.com>