ping: add tests
There was created a `TestSocket` which emulates socket.
It allows to monipulate the behaviour by its state.
Might it worth to be refactored in order to simplicity with some kind of
enum like `Side` and one map like `HashMap<Side, (usize, Action)>`.

```rust
enum Side {
  Send,
  Recv,
}
```

```rust
enum Action {
  Error(io::Error),
  Callback(fn(&mut IcmpBuilder)),
}
```

The test cannot be run in basic mode. There should be specified `--test-threads=1` flag.
I didn't come to the idea how to solve it yet.

Signed-off-by: Maxim Zhiburt <zhiburt@gmail.com>